+++ sampler/src/main/java/atlantafx/sampler/Resources.java
@@ -5,6 +5,7 @@ import java.io.InputStream;
 import java.net.URI;
 import java.net.URL;
 import java.util.Objects;
+import java.util.prefs.Preferences;
 
 public final class Resources {
 
@@ -31,5 +32,9 @@ public final class Resources {
 
     public static String getPropertyOrEnv(String propertyKey, String envKey) {
         return System.getProperty(propertyKey, System.getenv(envKey));
+    }
+
+    public static Preferences getPreferences() {
+        return Preferences.userRoot().node("atlantafx");
     }
 }