+++ sampler/src/main/java/module-info.java
@@ -5,6 +5,7 @@ module atlantafx.sampler {
     requires atlantafx.base;
 
     requires java.desktop;
+    requires java.prefs;
     requires javafx.swing;
     requires javafx.media;
     requires javafx.web;