+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## [Unreleased]
 
+### Features
+
+- (Sampler) External themes support. Sampler can now be used to develop custom themes. Just clone [AtlantaFX sample theme](https://github.com/mkpaz/atlantafx-sample-theme) repository, compile it and add resulting CSS file to the Sampler.
+- (CSS) `TextField` and `Button` now support rounded style.
+
 ### Improvements
 
-- `Hyperlink`, `TextField`, `TextArea` and `ProgressBar` can now be styled via looked-up color variables.
+- (CSS) Nord light and dark themes revamped with better color contrast and improved design.
+- (CSS) `Hyperlink`, `TextField`, `TextArea` and `ProgressBar` can now be styled via looked-up color variables.
+
+### Fixes
 
 ## [0.1]